Move BasicMap and SearchRequestMap to parent package
